+++ packages/api/test/integration/queries.test.ts
@@ -22,7 +22,7 @@ import {
   pageTypeOfficeDesksFetch,
   pageTypeOfficeFetch,
 } from '../mocks/CollectionQuery'
-import { ProductByIdQuery, productSearchFetch } from '../mocks/ProductQuery'
+import { ProductByIdQuery, pdpFetch } from '../mocks/ProductQuery'
 import {
   RedirectQueryTermTech,
   redirectTermTechFetch,
@@ -141,7 +141,7 @@ test('`collection` query', async () => {
 
 test('`product` query', async () => {
   const run = await createRunner()
-  const fetchAPICalls = [productSearchFetch, salesChannelStaleFetch]
+  const fetchAPICalls = [pdpFetch, salesChannelStaleFetch]
 
   mockedFetch.mockImplementation((info, init) =>
     pickFetchAPICallResult(info, init, fetchAPICalls)