+++ packages/core/src/components/account/components/MyAccountAccordion/index.ts
@@ -1,4 +1,0 @@
-export { default as MyAccountAccordion } from './MyAccountAccordion'
-export { default as MyAccountAccordionButton } from './MyAccountAccordionButton'
-export { default as MyAccountAccordionItem } from './MyAccountAccordionItem'
-export { default as MyAccountAccordionPanel } from './MyAccountAccordionPanel'