+++ packages/core/src/components/account/orders/MyAccountListOrders/MyAccountFilterSlider/MyAccountFilterFacetPendingApproval/index.ts
@@ -1,2 +1,0 @@
-export { default } from './MyAccountFilterFacetPendingApproval'
-export type { MyAccountFilterFacetPendingApprovalProps } from './MyAccountFilterFacetPendingApproval'