+++ packages/core/src/components/account/orders/MyAccountListOrders/index.ts
@@ -1,1 +1,0 @@
-export { default as MyAccountListOrders } from './MyAccountListOrders'