+++ packages/core/src/instrumentation.ts
@@ -1,22 +1,27 @@
-import config from '../discovery.config'
-import pkgJSON from '../package.json'
+/**
+ * Disabling telemetry instantiation
+ */
 
-export async function register() {
-  if (process.env.NEXT_RUNTIME === 'nodejs') {
-    const { name, version } = pkgJSON
-    try {
-      const { getTelemetryClient } = await import('@faststore/diagnostics')
-      console.log('Instrumentation.ts: Getting telemetry client')
+// import config from '../discovery.config'
+// import pkgJSON from '../package.json'
 
-      return getTelemetryClient({
-        serviceName: config.analytics?.serviceName ?? name,
-        version,
-        account: config.api.storeId,
-        clientName: config.api.storeId,
-        packageName: name,
-      })
-    } catch (error) {
-      console.error('Failed to initialize OTEL Instrumentation')
-    }
-  }
-}
+// export async function register() {
+//   if (process.env.NEXT_RUNTIME === 'nodejs') {
+//     const { name, version } = pkgJSON
+//     try {
+//       const { getTelemetryClient } = await import('@faststore/diagnostics')
+//       console.log('Instrumentation.ts: Getting telemetry client')
+
+//       return getTelemetryClient({
+//         serviceName: config.analytics?.serviceName ?? name,
+//         version,
+//         account: config.api.storeId,
+//         clientName: config.api.storeId,
+//         packageName: name,
+//       })
+//     } catch (error) {
+//       console.error('Failed to initialize OTEL Instrumentation')
+//     }
+//   }
+// }
+