+++ packages/core/src/pages/pvt/account/[...unknown].tsx
@@ -2,6 +2,7 @@ import type { Locator } from '@vtex/client-cms'
 import type { GetStaticPaths, GetStaticProps } from 'next'
 
 import type { MyAccountProps } from 'src/experimental/myAccountServerSideProps'
+import { localizeRedirectDestination } from 'src/utils/localization/localizeRedirectDestination'
 import { getMyAccountRedirect } from 'src/utils/myAccountRedirect'
 
 export default function Page() {
@@ -12,7 +13,7 @@ export const getStaticProps: GetStaticProps<
   MyAccountProps,
   Record<string, string>,
   Locator
-> = async () => {
+> = async ({ locale, defaultLocale }) => {
   const { isFaststoreMyAccountEnabled, redirect } = getMyAccountRedirect({
     query: {},
   })
@@ -23,7 +24,10 @@ export const getStaticProps: GetStaticProps<
 
   return {
     redirect: {
-      destination: '/pvt/account/404',
+      destination: localizeRedirectDestination('/pvt/account/404', {
+        locale,
+        defaultLocale,
+      }),
       permanent: false,
     },
   }